+++ src/main/java/localgaji/albbaim/schedule/application/DTO/ApplicationRequest.java
@@ -4,8 +4,6 @@ import io.swagger.v3.oas.annotations.media.Schema;
 
 import java.util.List;
 
-import static localgaji.albbaim.schedule.application.DTO.ApplicationDTO.*;
-
 public class ApplicationRequest {
     @Schema(description = "알바 스케줄 신청")
     public record PostApplyRequest(
@@ -14,5 +12,10 @@ public class ApplicationRequest {
             @Schema(description = "신청 정보")
             List<List<HasWorkTimeChecked>> apply
     ) {
+        public record HasWorkTimeChecked (
+                Long workTimeId,
+                Boolean isChecked
+        ) {
+        }
     }
 }
